@@ -1,17 +1,17 @@
-﻿namespace CloudyWing.FormValidators {
-    /// <summary>The form validator.</summary>
-    public interface IFormValidator {
-        /// <summary>Gets the error message.</summary>
-        /// <value>The error message.</value>
-        string ErrorMessage { get; }
+﻿namespace CloudyWing.FormValidators;
 
-        /// <summary>Returns true if ... is valid.</summary>
-        /// <value>
-        ///   <c>true</c> if this instance is valid; otherwise, <c>false</c>.</value>
-        bool IsValid { get; }
+/// <summary>The form validator.</summary>
+public interface IFormValidator {
+    /// <summary>Gets the error message.</summary>
+    /// <value>The error message.</value>
+    string ErrorMessage { get; }
 
-        /// <summary>Validates this instance.</summary>
-        /// <returns>The validation result.</returns>
-        bool Validate();
-    }
+    /// <summary>Returns true if ... is valid.</summary>
+    /// <value>
+    ///   <c>true</c> if this instance is valid; otherwise, <c>false</c>.</value>
+    bool IsValid { get; }
+
+    /// <summary>Validates this instance.</summary>
+    /// <returns>The validation result.</returns>
+    bool Validate();
 }
