@@ -1,43 +1,42 @@
 ﻿using CloudyWing.FormValidators.Core;
 using NUnit.Framework;
 
+namespace CloudyWing.FormValidators.Tests;
 
-namespace CloudyWing.FormValidators.Tests {
-    [TestFixture]
-    public class RequiredValidatorTests {
-        [TestCase("0", true)]
-        [TestCase("1", true)]
-        [TestCase(null, false)]
-        [TestCase("", false)]
-        [TestCase(" ", false)]
-        public void Validate_ReturnValue_AreEqual(string value, bool isValid) {
-            RequiredValidator validator = new RequiredValidator("", value);
+[TestFixture]
+public class RequiredValidatorTests {
+    [TestCase("0", true)]
+    [TestCase("1", true)]
+    [TestCase(null, false)]
+    [TestCase("", false)]
+    [TestCase(" ", false)]
+    public void Validate_ReturnValue_AreEqual(string value, bool isValid) {
+        RequiredValidator validator = new RequiredValidator("", value);
 
-            Assert.That(validator.Validate(), Is.EqualTo(isValid));
-        }
+        Assert.That(validator.Validate(), Is.EqualTo(isValid));
+    }
 
-        [Test]
-        public void ErrorMessage_DefaultMessage_AreEqual() {
-            string column = "測試欄位";
-            string value = null;
-            string expected = ErrorMessageProvider.ValueIsRequiredAccessor(column, value);
+    [Test]
+    public void ErrorMessage_DefaultMessage_AreEqual() {
+        string column = "測試欄位";
+        string value = null;
+        string expected = ErrorMessageProvider.ValueIsRequiredAccessor(column, value);
 
-            RequiredValidator validator = new RequiredValidator(column, value);
-            validator.Validate();
+        RequiredValidator validator = new RequiredValidator(column, value);
+        validator.Validate();
 
-            Assert.That(validator.ErrorMessage, Is.EqualTo(expected));
-        }
+        Assert.That(validator.ErrorMessage, Is.EqualTo(expected));
+    }
 
-        [Test]
-        public void ErrorMessage_CustomMessage_AreEqual() {
-            string column = "測試欄位";
-            string value = null;
-            string expected = column + value;
+    [Test]
+    public void ErrorMessage_CustomMessage_AreEqual() {
+        string column = "測試欄位";
+        string value = null;
+        string expected = column + value;
 
-            RequiredValidator validator = new RequiredValidator(column, null, (c, v) => c + v);
-            validator.Validate();
+        RequiredValidator validator = new RequiredValidator(column, null, (c, v) => c + v);
+        validator.Validate();
 
-            Assert.That(validator.ErrorMessage, Is.EqualTo(expected));
-        }
+        Assert.That(validator.ErrorMessage, Is.EqualTo(expected));
     }
 }
